+++ lib/views/questions_view.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_ecoapp/models/article.dart';
 import 'package:flutter_ecoapp/models/question.dart';
 import 'package:flutter_ecoapp/views/style/colors.dart';
-import 'package:flutter_ecoapp/views/widgets/articles/article_cover.dart';
 import 'package:flutter_ecoapp/views/widgets/bottom_nav_bar.dart';
+import 'package:flutter_ecoapp/views/widgets/eco_cover.dart';
 import 'package:flutter_ecoapp/views/widgets/eco_items_list.dart';
 import 'package:google_fonts/google_fonts.dart';
 
@@ -54,9 +54,11 @@ class QuestionsView extends StatelessWidget {
         ),
         child: Column(
           children: [
-            ArticleCover(
-              article: article,
+            EcoCover(
+              image: NetworkImage(article.photos[0].photoUrl),
+              title: article.title,
               subtitle: '${article.questions.length.toString()} preguntas',
+              size: 80
             ),
             SizedBox(height: 10.0,),
             Divider(thickness: 1,),
