+++ lib/views/widgets/store/storeview/store_cover_section.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_ecoapp/models/store.dart';
 import 'package:flutter_ecoapp/views/store_rating_view.dart';
+import 'package:flutter_ecoapp/views/widgets/eco_cover.dart';
 import 'package:flutter_ecoapp/views/widgets/stars_row.dart';
 import 'package:google_fonts/google_fonts.dart';
 
@@ -14,64 +15,6 @@ class StoreCoverSection extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final image = Hero(
-      tag: store.tag,
-      child: Container(
-        width: 100,
-        height: 100,
-        clipBehavior: Clip.antiAlias,
-        decoration: BoxDecoration(
-          shape: BoxShape.circle,
-          boxShadow: [
-            BoxShadow(
-              color: Colors.black.withOpacity(0.25),
-              blurRadius: 5
-            )
-          ]
-        ),
-        child: Image(
-          image: NetworkImage(store.photoUrl),
-          fit: BoxFit.cover,
-        ),
-      ),
-    );
-
-    final info = Container(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.end,
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        children: [
-          Text(
-            '${store.location}, ${store.district.name}',
-            style: GoogleFonts.montserrat(
-              fontWeight: FontWeight.w500
-            ),
-            textAlign: TextAlign.end,
-          ),
-          SizedBox(height: 10.0),
-          Text(
-            'Ventas concretadas: 17.350',
-            style: GoogleFonts.montserrat(),
-            textAlign: TextAlign.end,
-          ),
-          SizedBox(height: 10.0),
-          Text(
-            'Creado el ${store.createdDate.day}/${store.createdDate.month}/${store.createdDate.year}',
-            style: GoogleFonts.montserrat(),
-            textAlign: TextAlign.end,
-          )
-        ],
-      ),
-    );
-
-    final firstRow = Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        image,
-        SizedBox(width: 20.0),
-        Expanded(child: info)
-      ],
-    );
 
     final description = Container(
       margin: EdgeInsets.symmetric(
@@ -131,7 +74,12 @@ class StoreCoverSection extends StatelessWidget {
       ),
       child: Column(
         children: [
-          firstRow,
+          EcoCover(
+            image: NetworkImage(store.photoUrl),
+            title: '${store.location}, ${store.district.name}',
+            subtitle: 'Ventas concretadas: 17.350',
+            miniContent: 'Creado el ${store.createdDate.day}/${store.createdDate.month}/${store.createdDate.year}',
+          ),
           SizedBox(height: 10.0),
           Divider(thickness: 1,),
           description,
